done with everything
added the 3rd and 4th section and also the footer. looks complete to me,
suggest me if I could do anything in a better way(i know there will be
many things since this is my first css project) and also I dont think
anyone is going to read this commit message but still just because I
wanna type and make the commit longer for no reason Im typing stuff
that even I dont think I will be reading anytime in the future yes but
one last thing, this project was challenging since I started from nothi
-ng and yet with the reference of the image I built the css layout of
the site as it is(hopefully) so it's a big achievement for me so
Im proud of myself and Im congratulating myself for the same. And I wish
to learn more and build more challenging stuff and end up becoming a
great developer!